+++ src/scripts/index.js
@@ -92,12 +92,12 @@ import UiTooltip from './components/others/tooltip';
 /**
  * Plugins
  */
-import event, { useBus } from './plugins/event'; // Custom
-import store from './plugins/store'; // Custom
-import grid, { useGrid } from './plugins/grid';
-import theme, { useTheme } from './plugins/theme';
-import typography, { useTypography } from './plugins/typography';
-import validator from './plugins/validator'; // Custom
+import $event, { useBus } from './plugins/event'; // Custom
+import $store from './plugins/store'; // Custom
+import $grid, { useGrid } from './plugins/grid';
+import $theme, { useTheme } from './plugins/theme';
+import $typography, { useTypography } from './plugins/typography';
+import $validator from './plugins/validator'; // Custom
 /**
  * Directives
  */
@@ -196,12 +196,12 @@ const components = {
 };
 
 const plugins = {
-  event,
-  grid,
-  store,
-  theme,
-  typography,
-  validator
+  $event,
+  $grid,
+  $store,
+  $theme,
+  $typography,
+  $validator
 };
 
 const directives = {
@@ -326,7 +326,7 @@ export {
   UiTooltip,
   UiTopAppBar
 };
-export { event, grid, store, theme, typography, validator };
+export { $event, $grid, $store, $theme, $typography, $validator };
 export { useBus, useGrid, useTheme, useTypography };
 export { vAccessibility, vBadge, vRipple, vShadow, vShape, vTooltip };
 export { types, helpers, detectIE };